Moved the bytecode-walker from evalfilter -> vm
This is the logical place for it to live, although I've also
got to move the optimizer - and to do that I have to cheat and
use the environment to pass the flag across packages (as I did
with the debug-handling).
