+++ evalfilter.go
@@ -8,7 +8,6 @@
 package evalfilter
 
 import (
-	"encoding/binary"
 	"fmt"
 	"strings"
 
@@ -19,15 +18,6 @@ import (
 	"github.com/skx/evalfilter/v2/parser"
 	"github.com/skx/evalfilter/v2/vm"
 )
-
-// BytecodeVisitor is the function-signature of the callbackup function which
-// can be invoked to interate over our generated bytecode.
-//
-// The callback function can be used via the `WalkBytecode` method, and the
-// return values should make sense:  If there is an error then that is returned
-// otherwise the bool-value will control whether the iteration continues,
-// return true to keep walking, and false to abort the process.
-type BytecodeVisitor func(offset int, instruction code.Opcode, argument interface{}) (bool, error)
 
 // Flags which can be optionally passed to Prepare.
 const (
@@ -131,14 +121,12 @@ func (e *Eval) Prepare(flags ...[]byte) error {
 	}
 
 	//
-	// Attempt to optimize the code, running multiple passes until no
-	// more changes are possible.
-	//
-	// We do this so that each optimizer run only has to try one thing
-	// at a time.
+	// If we've got the optimizer enabled then set the environment
+	// variable, so that the virtual machine knows it should
+	// run a series of optimizations.
 	//
 	if optimize {
-		e.optimize()
+		e.environment.Set("OPTIMIZE", &object.Boolean{Value: true})
 	}
 
 	//
@@ -146,6 +134,9 @@ func (e *Eval) Prepare(flags ...[]byte) error {
 	// we've created - as well as any function pointers and variables
 	// which we were given.
 	//
+	// The optimization will happen at this step, so that it is complete
+	// before Execute/Run are invoked - and we only take the speed hit
+	// once.
 	e.machine = vm.New(e.constants, e.instructions, e.environment)
 
 	//
@@ -154,29 +145,28 @@ func (e *Eval) Prepare(flags ...[]byte) error {
 	return nil
 }
 
-// Dump causes our bytecode to be dumped.
-//
-// This is used by the `evalfilter` CLI-utility, but it might be useful
-// to consumers of our library.
+// Dump causes our bytecode to be dumped, along with the contents
+// of the constant-pool
 func (e *Eval) Dump() error {
 
 	fmt.Printf("Bytecode:\n")
 
-	// Use the walker to dump our bytecode.
-	e.WalkBytecode(func(offset int, opCode code.Opcode, opArg interface{}) (bool, error) {
+	// Use the walker to dump the bytecode.
+	e.machine.WalkBytecode(func(offset int, opCode code.Opcode, opArg interface{}) (bool, error) {
 
 		// Show the offset + instruction.
-		fmt.Printf("%06d\t%14s", offset, code.String(opCode))
+		fmt.Printf("%04d\t%14s", offset, code.String(opCode))
 
+		// Show the optional argument, if present.
 		if opArg != nil {
-			fmt.Printf("\t%d", opArg.(int))
+			fmt.Printf("\t% 4d", opArg.(int))
 		}
 
 		// Some opcodes benefit from inline comments
 		if code.Opcode(opCode) == code.OpConstant {
 			v := e.constants[opArg.(int)]
 			s := strings.ReplaceAll(v.Inspect(), "\n", "\\n")
-			fmt.Printf("\t// load constant: \"%s\"", s)
+			fmt.Printf("\t// push constant onto stack: \"%s\"", s)
 		}
 		if code.Opcode(opCode) == code.OpLookup {
 			fmt.Printf("\t// lookup field/variable: %v", e.constants[opArg.(int)])
@@ -188,17 +178,17 @@ func (e *Eval) Dump() error {
 			fmt.Printf("\t// Push %d to stack", opArg.(int))
 		}
 		fmt.Printf("\n")
+
+		// Keep walking, no error.
 		return true, nil
 	})
 
 	// Show constants, if any are present.
 	if len(e.constants) > 0 {
-		fmt.Printf("\n\nConstants:\n")
+		fmt.Printf("\n\nConstant Pool:\n")
 		for i, n := range e.constants {
-
 			s := strings.ReplaceAll(n.Inspect(), "\n", "\\n")
-
-			fmt.Printf("  %06d Type:%s Value:\"%s\"\n", i, n.Type(), s)
+			fmt.Printf("%04d Type:%s Value:\"%s\"\n", i, n.Type(), s)
 		}
 	}
 
@@ -288,72 +278,4 @@ func (e *Eval) GetVariable(name string) object.Object {
 		return value
 	}
 	return &object.Null{}
-}
-
-// WalkBytecode invokes the specified callbackup function upon every
-// instruction in our generated bytecode.
-//
-// This is used by our dumping command, as well as by the optimizer.
-func (e *Eval) WalkBytecode(callback BytecodeVisitor) error {
-
-	//
-	// We're going to walk over the bytecode from start to
-	// finish.
-	//
-	ip := 0
-	ln := len(e.instructions)
-
-	//
-	// Walk the bytecode.
-	//
-	for ip < ln {
-
-		//
-		// Get the next opcode
-		//
-		op := code.Opcode(e.instructions[ip])
-
-		//
-		// Find out how long it is.
-		//
-		opLen := code.Length(op)
-
-		//
-		// If the opcode is more than a single byte long
-		// we read the argument here.
-		//
-		opArg := 0
-		if opLen > 1 {
-
-			//
-			// Note in the future we might have to cope
-			// with opcodes with more than a single argument,
-			// and they might be different sizes.
-			//
-			opArg = int(binary.BigEndian.Uint16(e.instructions[ip+1 : ip+3]))
-		}
-
-		var err error
-		var ret bool
-
-		if opLen > 1 {
-			ret, err = callback(ip, op, opArg)
-		} else {
-			ret, err = callback(ip, op, nil)
-		}
-
-		// Error?  Then return that, and stop walking.
-		if err != nil {
-			return err
-		}
-
-		// The callback returned false?  Then stop walking.
-		if !ret {
-			return nil
-		}
-
-		ip += opLen
-	}
-
-	return nil
 }
