+++ vm/vm.go
@@ -22,6 +22,15 @@ import (
 	"github.com/skx/evalfilter/v2/stack"
 )
 
+// BytecodeVisitor is the function-signature of the callbackup function which
+// can be invoked to interate over our generated bytecode.
+//
+// The callback function can be used via the `WalkBytecode` method, and the
+// return values should make sense:  If there is an error then that is returned
+// otherwise the bool-value will control whether the iteration continues,
+// return true to keep walking, and false to abort the process.
+type BytecodeVisitor func(offset int, instruction code.Opcode, argument interface{}) (bool, error)
+
 // True is our global "true" object.
 var True = &object.Boolean{Value: true}
 
@@ -68,17 +77,35 @@ type VM struct {
 }
 
 // New constructs a new virtual machine.
+//
+// If the value `OPTIMIZE` exists inside the environment we're passed we'll also run a
+// series of simple optimizer steps.  These are naive, but do speedup carefully constructed
+// test cases.
 func New(constants []object.Object, bytecode code.Instructions, env *environment.Environment) *VM {
 
-	// If we have a `DEBUG` environment then we enable debugging
-	_, present := env.Get("DEBUG")
+	// If we have a `DEBUG` environment then we enable debugging.
+	_, debug := env.Get("DEBUG")
 
-	return &VM{
+	// If we have "OPTIMIZE" set then we optimize our bytecode.
+	//
+	// We don't need to store this flag anywhere, we'll just run
+	// the optimizer immediately.
+	_, optimize := env.Get("OPTIMIZE")
+
+	// Create the machine
+	vm := &VM{
 		constants:   constants,
 		environment: env,
 		bytecode:    bytecode,
-		debug:       present,
+		debug:       debug,
 	}
+
+	// Optimize the bytecode, if we should.
+	if optimize {
+		vm.optimizeBytecode()
+	}
+
+	return vm
 }
 
 // Run launches our virtual machine, intepreting the bytecode-program we were
@@ -92,7 +119,9 @@ func New(constants []object.Object, bytecode code.Instructions, env *environment
 // a hand-created program could build such a things via the instruction-set.)
 func (vm *VM) Run(obj interface{}) (object.Object, error) {
 
+	//
 	// Sanity-check the bytecode program is non-empty
+	//
 	if len(vm.bytecode) < 1 {
 		return nil, fmt.Errorf("the bytecode program is empty")
 	}
@@ -132,7 +161,14 @@ func (vm *VM) Run(obj interface{}) (object.Object, error) {
 	// Loop over all the bytecode.
 	//
 	// Note that the instruction set supports control-flow, so it
-	// is possible we'll run forever..
+	// is possible we'll run forever.
+	//
+	// NOTE: We should port this to use our WalkBytecode function,
+	// however we've not done so because that won't cope with changing
+	// the instruction-pointer, which we need to implement the jumping
+	// primitives.
+	//
+	// This is an area that needs some thought.
 	//
 	for ip < ln {
 
@@ -577,6 +613,7 @@ func (vm *VM) createArrayFromSlice(field reflect.Value) object.Object {
 			continue
 		}
 
+		// Is it a time value?
 		tm, ok := in.(time.Time)
 		if ok {
 			el = append(el, &object.Integer{Value: tm.Unix()})
@@ -1054,5 +1091,75 @@ func (vm *VM) executeIndexExpression(left, index object.Object) error {
 
 	// Return the appropriate object.
 	vm.stack.Push(arrayObject.Elements[idx])
+	return nil
+}
+
+// WalkBytecode invokes the specified callbackup function upon every
+// instruction in our bytecode program.
+//
+// This is primarily used for our optimizer, but it is also used
+// for the implementation of the Dump command in our evalfilter
+// package.
+func (vm *VM) WalkBytecode(callback BytecodeVisitor) error {
+
+	//
+	// We're going to walk over the bytecode from start to finish.
+	//
+	ip := 0
+	ln := len(vm.bytecode)
+
+	//
+	// Walk the bytecode.
+	//
+	for ip < ln {
+
+		//
+		// Get the next opcode
+		//
+		op := code.Opcode(vm.bytecode[ip])
+
+		//
+		// Find out how long it is.
+		//
+		opLen := code.Length(op)
+
+		//
+		// If the opcode is more than a single byte long we read the argument here.
+		//
+		opArg := 0
+		if opLen > 1 {
+
+			//
+			// Note in the future we might have to cope
+			// with opcodes with more than a single argument,
+			// and they might be different sizes.
+			//
+			opArg = int(binary.BigEndian.Uint16(vm.bytecode[ip+1 : ip+3]))
+		}
+
+		var err error
+		var ret bool
+
+		// Pass either the argument, or nil.
+		if opLen > 1 {
+			ret, err = callback(ip, op, opArg)
+		} else {
+			ret, err = callback(ip, op, nil)
+		}
+
+		// Error?  Then return that, and stop walking.
+		if err != nil {
+			return err
+		}
+
+		// The callback returned false?  Then stop walking.
+		if !ret {
+			return nil
+		}
+
+		ip += opLen
+	}
+
+	// No error, walk is complete.
 	return nil
 }
